requirements-voice.txt
# W.I.T. Voice System Requirements
# COPY FROM ARTIFACT: wit-voice-requirements

# Core requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Voice processing (optional but recommended)
SpeechRecognition>=3.10.0
pyaudio>=0.2.13

# Audio processing
numpy>=1.24.0
scipy>=1.11.0

# For development/testing
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
